First implemenation of the factory method with the new controller.
