+++ zaber/CMakeLists.txt
@@ -2,10 +2,10 @@
 set(TARGET_LIB_NAME ${CMAKE_PROJECT_NAME}_lib)
 
 add_library(${TARGET_LIB_NAME}
-  include/zaber/zaber_types.h
-  include/zaber/zaber.h
-  include/zaber/zaber_ascii.h
-  src/zaber.cpp
+  src/zaber_control.cpp
+  src/zaber_device.cpp
+  src/zaber_types.cpp
+  src/zaber_rotary.cpp
  )
 
 target_include_directories(${TARGET_LIB_NAME} PUBLIC
