--- zaber/include/zaber/X_RSW_E.h
@@ -1,33 +1,0 @@
-/*
-* @Author: Lukasz
-* @Date:   02-10-2018
-* @Last Modified by:   Lukasz
-* @Last Modified time: 03-10-2018
-@brief      Zaber's X-RSW-E Series products are computer controlled motorized
-            rotary stages with integrated controllers and motor encoders. They
-            are stand-alone units requiring only a standard 24-48 V power
-            supply. Built-in motor encoders allow for closed-loop operation and
-            slip/stall recovery features. An indexed knob provides smooth manual
-            control at variable speeds in both directions for versatile
-            operation even without a computer. Press and hold to switch between
-            velocity mode and distance mode, turn to move the stage, and press
-            to stop. The stages connect to the USB 2.0 or RS-232 port of any
-            computer and can be chained with several units per chain. They can
-            also be chained with any other Zaber products. Convenient locking,
-            4-pin, M8 connectors on the unit allow for easy and secure
-            connection between products. The chain also shares power, so
-            multiple X-Series products can use a single power supply.
-*/
-
-#pragma once
-
-#include "zaber/zaber.h"
-
-namespace zaber
-{
-
-// class DeviceRotary : public Device
-// {
-// };
-
-} // namespace zaber
