--- zaber/include/zaber/zaber_comm.h
@@ -1,17 +1,0 @@
-/*
- * @Author: Lukasz
- * @Date:   02-10-2018
- * @Last Modified by:   Lukasz
- * @Last Modified time: 02-10-2018
- */
-#pragma once
-
-namespace zaber
-{
-class CommPort
-{
-public:
-  CommPort();
-  ~CommPort();
-};
-} // namespace zaber
