+++ zaber/include/zaber/zaber_includes.h
@@ -1,0 +1,29 @@
+/*
+* @Author: Lukasz
+* @Date:   04-10-2018
+* @Last Modified by:   Lukasz
+* @Last Modified time: 04-10-2018
+*/
+
+#pragma once
+
+#include "fmt/format.h"
+#include "fmt/ostream.h"
+#include "serial/serial.h"
+#include "spdlog/sinks/stdout_color_sinks.h"
+#include <cstdio>
+#include <exception>
+#include <iostream>
+#include <memory> //Shared ptr.
+#include <spdlog/spdlog.h>
+#include <stdint.h>
+#include <string>
+#include <unistd.h>
+
+#include <iostream>
+#include <math.h> /* log10 */
+#include <stdint.h>
+#include <stdio.h> /* printf */
+#include <stdlib.h>
+#include <string>
+#include <vector>
